create list item test works
