+++ index.html
@@ -11,13 +11,7 @@
     <header>
         <h1>Filtered Image Library</h1>
     </header>
-    <ul id="photo-gallery">
-        <li>
-            <h2>title</h2>
-            <img src="http://3.bp.blogspot.com/_DBYF1AdFaHw/TE-f0cDQ24I/AAAAAAAACZg/l-FdTZ6M7z8/s1600/Unicorn_and_Narwhal_by_dinglehopper.jpg">
-            <p>horns: 1</p>
-        </li>
-    </ul>
-    
+    <ul id="photo-gallery"></ul>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
