+++ main.css
@@ -3,6 +3,7 @@ body h1{
 }
 
 #photo-gallery li{
+    float:right;
     display:flex;
     flex-direction:column;
     align-items:center;
@@ -10,6 +11,7 @@ body h1{
     width:200px;
     list-style: none;
     border: 2px solid red; 
+    margin:5px;
 }
 
 #photo-gallery img {
